requirements.txt
torch
torchvision
diffusers
transformers
accelerate
gradio
pillow
